// src/FE/src/api/generated/models/createJobWorkRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { CreateInstallationTypeRequest } from './createInstallationTypeRequest';

export interface CreateJobWorkRequest {
  /** @nullable */
  installationTypes: CreateInstallationTypeRequest[] | null;
  /** @nullable */
  workKind: string | null;
  /** @nullable */
  customWorkKind: string | null;
  /** @nullable */
  closureFlags: string[] | null;
  /** @nullable */
  remarks: string | null;
}